userProfileCrud

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
   try {
@@ -12,20 +13,31 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id); // id를 통해 video를 검색
+  const video = await Video.findById(id).populate("owner");
+  // id를 통해 video를 검색
+  // populate를 통해 mongoose가 User의 정보를 "owner"에 입력해 줌,
+
   if (!video) {
     res.status(404).render("404", { pageTitle: "Video not found" }); // video가 없을시 404 렌더링
   } else {
-    res.render("watch", { pageTitle: video.title, video });
+    res.render("videos/watch", { pageTitle: video.title, video });
   }
 };
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const { _id } = req.session.user;
   const video = await Video.findById(id); // id를 통해 video를 검색
+
   if (!video) {
     res.render("404", { pageTitle: "Video not found" }); // video가 없을시 404 렌더링
   }
+
+  // 만약 video의 owner가 아닌 사람이 edit창을 열려고 하는 것을 방지/=.
+  if (String(video.owner) !== _id) {
+    return res.status(403).redirect("/");
+  }
+
   return res.render("videos/edit", { pageTitle: `Edit ${video.title}`, video });
 };
 
@@ -50,19 +62,24 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
-  // here we will add a video
+  const {
+    user: { _id },
+  } = req.session;
   const file = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
-      // Video 모델을 이용하여 video 제작
+    const newVideo = await Video.create({
+      // Video 모델을 이용하여 video 제작, newVideo는 제작된 video값을 return
       title,
       description,
       createdAt: Date.now(),
       fileUrl: file.path,
-      hashtags: Video.formatHashtags(hashtags),
-      // Video model에서 만든 formatHashtags 함수를 import
+      hashtags: Video.formatHashtags(hashtags), // Video model에서 만든 formatHashtags 함수를 import
+      owner: _id, // video의 owner 파악
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id); // User모델에 있는 배열에 _id를 push한다.
+    user.save();
     res.redirect("/");
   } catch (error) {
     console.log(error);
@@ -75,7 +92,22 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const { _id } = req.session.user;
+
+  const video = await Video.findById(id); // id를 통해 video를 검색
+
+  if (!video) {
+    res.render("404", { pageTitle: "Video not found" }); // video가 없을시 404 렌더링
+  }
+
+  // 만약 video의 owner가 아닌 사람이 edit창을 열려고 하는 것을 방지/=.
+  if (String(video.owner) !== _id) {
+    console.log("fail");
+    return res.status(403).redirect("/");
+  }
+
   await Video.findByIdAndDelete(id);
+
   return res.redirect("/");
 };
